remove subshells, use `-q`

+++ examples/projects/mongocxx/pkg-config/static/build.sh
@@ -5,24 +5,21 @@ set -o pipefail
 
 # Sanity-check that static library macros are set when building against the static library.  Users
 # don't need to include this section in their projects.
-(pkg-config --cflags libmongocxx-static | grep -- -DBSONCXX_STATIC) ||
-  (
-    echo "Expected BSONCXX_STATIC to be set" >&2
-    exit 1
-  )
-(pkg-config --cflags libmongocxx-static | grep -- -DMONGOCXX_STATIC) ||
-  (
-    echo "Expected MONGOCXX_STATIC to be set" >&2
-    exit 1
-  )
+if ! pkg-config --cflags libmongocxx-static | grep -q -- -DBSONCXX_STATIC; then
+  echo "Expected BSONCXX_STATIC to be set" >&2
+  exit 1
+fi
+
+if ! pkg-config --cflags libmongocxx-static | grep -q -- -DMONGOCXX_STATIC; then
+  echo "Expected MONGOCXX_STATIC to be set" >&2
+  exit 1
+fi
 
 # Sanity-check that static libmongoc is required. Regression test for CXX-3290.
-(pkg-config --print-requires libmongocxx-static | grep -- mongoc2-static || (
-  (
-    echo "Expected mongoc2-static to be required" >&2
-    exit 1
-  )
-))
+if ! pkg-config --print-requires libmongocxx-static | grep -q -- mongoc2-static; then
+  echo "Expected mongoc2-static to be required" >&2
+  exit 1
+fi
 
 rm -rf build/*
 cd build
